# JBR-7392: Use NIO FS in ZipFile