See https://youtrack.jetbrains.com/issue/JBR-7392/Allow-using-NIO-in-ZipFile for discussion.

Tests against upstream JDK: https://github.com/knisht/jdk/pull/1
Tests with enabled property: https://github.com/knisht/jdk/pull/2. It contains an additional failure of `RedefineSharedClass`, which is fixable.