@mkartashev 2024-07-09
> It contains an additional failure of RedefineSharedClass, which is fixable.

Please, provide a detailed explanation of the failure and how it can be fixed.