@mkartashev 2024-07-09
> Since ZipFile is used in VM startup, it means that the VM loses the ability to work without allocation of direct buffers

Can you make the "new" ZipFile parametrized by the VM bootup stage so that it can be successfully used both during the JVM startup (in the RandomAccessFile mode) and later (in the NIO mode)?