@mkartashev 2024-07-09
What are performance implications outside of WSL? "Regular" Windows, macOS, Linux? How does this patch affect startup time?