@knisht 2024-07-09
> Please, provide a detailed explanation of the failure and how it can be fixed.

`RedefineSharedClass` fails on the following line:
```java
        // When run without arguments this class acts as the test. First redefining
        // a class that we expect to be in the archive if used and the triggering a
        // System.gc() to clean up.
        RedefineClassHelper.redefineClass(java.io.RandomAccessFile.class, getClassBytes(java.io.RandomAccessFile.class));
```
Since `java.io.RandomAccessFile` is not loaded in `ZipFile` anymore, we cannot expect that it appears in the archive. If we replace `java.io.RandomAccessFile` with `java.util.zip.ZipFile`, the test passes, and it checks what it used to check.

There is also another failure in `TestZipError`, and it is Windows-related.
```java
        // Open the ZipFile.  This will read the zip file's central
        // directory into in-memory data structures.
        ZipFile zf = new ZipFile(fileName);

        // Delete the file; of course this does not change the in-memory data
        // structures that represent the central directory!
        f.delete();

        // Re-create zip file, with different entries than earlier.  However,
        // recall that we have in-memory information about the central
        // directory of the file at its previous state.
        zos = new ZipOutputStream(new FileOutputStream(f));
```
The test fails here with `FileNotFoundException`. 
The reason is that `RandomAccessFile` opens a file without `FILE_SHARE_DELETE` (https://github.com/openjdk/jdk/blob/7e11fb702696df733ca89d325200f2e9414402d9/src/java.base/windows/native/libjava/io_util_md.c#L220), while `FileChannel` has this share mode enabled by default (https://github.com/openjdk/jdk/blob/7e11fb702696df733ca89d325200f2e9414402d9/src/java.base/windows/classes/sun/nio/fs/WindowsChannelFactory.java#L121). 
In the case of `RandomAccessFile`, the delete fails (so the test's behavior is not really meaningful), and in the case of `FileChannel` we get a well-deserved exception. This can be fixed simply by removing `f.delete()`.