@knisht 2024-07-09
> Can you make the "new" ZipFile parametrized by the VM bootup stage so that it can be successfully used both during the JVM startup (in the RandomAccessFile mode) and later (in the NIO mode)?

What are we trying to achieve here? Whenever the VM accesses a jar file (for lazy classloading, as an example), it would allocate a memory-mapped file. I don't think it is possible to have `FileChannel`-based `ZipFile` without the use of direct memory in VM's internal operations.