@mkartashev 2024-07-10
> > Can you make the "new" ZipFile parametrized by the VM bootup stage so that it can be successfully used both during the JVM startup (in the RandomAccessFile mode) and later (in the NIO mode)?
> 
> What are we trying to achieve here? 

Never mind, I thought this is what was causing the test to fail.