@mkartashev 2024-07-10
Thanks for explaining the failures. Could you, please, fix those two tests so that we don't have to remember to exclude them forever?