@mkartashev 2024-07-10
BTW, those tests need to _also_ be run with `-Djava.util.zip.use.nio.for.zip.file.access=true`