@knisht 2024-07-12
## Benchmark description
I made a small performance test which loads all classes from the provided class path. 
1. The test itself uses an inheritor of `ZipFile`.
2. The test loads classes, and classloader is a client of `ZipFile`.
```java
import java.io.File;
import java.io.IOException;
import java.util.jar.JarFile;

public class Main {

  public static void main(String[] args) throws IOException {
    ClassLoader cl = Thread.currentThread().getContextClassLoader();

    String cp = System.getProperty("java.class.path");
    String[] jars = cp.split(File.pathSeparator);
    for (String jar : jars) {
      try (JarFile jarFile = new JarFile(jar)) {
        jarFile.stream().forEach(entry -> {
          if (entry.getName().endsWith(".class")) {
            String className = entry.getName().replace('/', '.').substring(0, entry.getName().length() - 6);
            try {
              Class.forName(className, true, cl);
            }
            catch (Throwable e) {
            }
          }
        });
      }
    }
    System.exit(0);
  }
}
```
I applied the test to a classpath consisting of several fat jars. For example, the archive with IDE classes takes around 120 MB of space, so I consider it a reasonable object of measurement. The test loaded 70935 classes in total.

Full command line looks like this: 
```sh
./java -Djava.util.zip.use.nio.for.zip.file.access=<true/false> -cp app-client.jar:kotlin-stdlib-2.0.0.jar:util-8.jar <benchmark file.java>
```

## Performance

I wrote a script which measures time of execution. The script invokes the command above 10 times, gathering some statistics about the execution time.

### MacOS (aarch64)
Without NIO in `ZipFile`:
```
Mean Execution Time: 8.703340 seconds
Standard Deviation of Execution Time: 0.594840 seconds
```
With NIO in `ZipFile`
```
Mean Execution Time: 8.694077 seconds
Standard Deviation of Execution Time: 0.389752 seconds
```
 ### Windows (x64)

 Without NIO:
 ```
Mean Execution Time: 11.088352 seconds
Standard Deviation of Execution Time: 0.645189 seconds
 ```
With NIO:
```
Mean Execution Time: 10.799217 seconds
Standard Deviation of Execution Time: 0.309649 seconds
```

### Linux (x64, inside WSL)

Without NIO:
```
Mean Execution Time: 9.980494 seconds
Standard Deviation of Execution Time: 0.254731 seconds
```
With NIO:
```
Mean Execution Time: 10.369276 seconds
Standard Deviation of Execution Time: 0.521237 seconds
```