@mkartashev 2024-07-12
Thank you, sounds reasonable. There doesn't seem to be any impact on startup time, so I will go ahead and port this change to `jbr21`.